pyproject.toml
[project]
name = "api-docker-auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.2.1",
    "dotenv-python>=0.0.1",
    "fastapi>=0.115.8",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.38",
    "telebot>=0.0.5",
    "uvicorn>=0.34.0",
]
